refactor: update background-jobs to have only SQSHandlers
