+++ services/background-jobs/handlers/demo.ts
@@ -1,0 +1,11 @@
+import { SQSHandler } from 'aws-lambda';
+
+export const handler: SQSHandler = async (event) => {
+  const recordHandler = async (record) => {
+    console.log(JSON.stringify(record.body));
+  };
+
+  // Ensuring we await on all the promises is super important to avoid
+  // accidentally killing the lambda prior to processing being completed.
+  await Promise.all(event.Records.map(recordHandler));
+};
